Web Inspector: dock buttons disappear if Web Inspector goes fullscreen https://bugs.webkit.org/show_bug.cgi?id=189865 <rdar://problem/42600534>
Reviewed by Matt Baker.

When a fullscreen Safari tab has a docked Inspector, and the "detach" button
is clicked, Inspector goes into its own fullscreen window. In that window,
there are no buttons for docking to side or bottom.

It turns out that we always send setDockingUnavailable(true) if the inspector
window is fullscreen. This eventually causes the inspector to reflect that
by hiding the docking-related buttons.

* UIProcess/mac/WebInspectorProxyMac.mm:
(WebKit::WebInspectorProxy::platformCanAttach):
This early exit does not seem to serve any purpose, so remove it.


Identifier: 204828@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@236375 268f45cc-cd09-0410-ab3c-d52691b4dbfc