RenderLayer::removeOnlyThisLayer() should not call updateLayerPositions() https://bugs.webkit.org/show_bug.cgi?id=190093
Reviewed by Dean Jackson and Zalan Bujtas.

Source/WebCore:

It's wrong for RenderLayer::removeOnlyThisLayer() to call updateLayerPositions(),
because this is called at style update time, and layout will be stale.

It was added (see webkit.org/b/25252) so that opacity changes, which can destroy layers, correctly update
descendants. However, RenderStyle::changeRequiresLayout() checks for opacity <=> no opacity
changes and triggers layout accordingly, which will result in a full post-layout
updateLayerPositions().

This also revealed that changes to the "isolate" property fail to trigger any kind of style recalc or layout;
we need it to trigger layout (for now) because it affects z-order.

Covered by existing tests.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::removeOnlyThisLayer):
* rendering/style/RenderStyle.cpp:
(WebCore::RenderStyle::changeRequiresLayout const):

LayoutTests:

The ordering of the repaints changes.

blend-mode-turn-off-isolation-no-effect.html now issues a repaint, which is expected now.
Isolation is so rare that this is not a problem.

* css3/blending/repaint/blend-mode-isolate-stacking-context-expected.txt:
* css3/blending/repaint/blend-mode-turn-off-isolation-expected.txt:
* css3/blending/repaint/blend-mode-turn-off-isolation-no-effect-expected.txt:
* fast/repaint/absolute-position-change-containing-block-expected.txt:


Identifier: 205062@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@236632 268f45cc-cd09-0410-ab3c-d52691b4dbfc