Clicking on bookmarklet crashes at NavigationState::willRecordNavigationSnapshot https://bugs.webkit.org/show_bug.cgi?id=190476 <rdar://problem/32757191>
Reviewed by Megan Gardner.

Source/WebKit:

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _saveBackForwardSnapshotForItem:]):
The WKBackForwardListItem can be nil. Then we retrieve a reference from it.
An Objective-C method that returns a reference is a great way to get
a null reference. Other code then assumes it's not null and crashes.
Add a nil check.

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/SnapshotStore.mm:
(TEST):
Add a test that we don't crash.


Identifier: 205429@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237055 268f45cc-cd09-0410-ab3c-d52691b4dbfc