Web Inspector: ⌃G to jump to line in CSS file wipes the line https://bugs.webkit.org/show_bug.cgi?id=190645
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2018-10-16
Reviewed by Devin Rousso.

* UserInterface/Base/Main.js:
* UserInterface/Views/Dialog.js:
(WI.Dialog.prototype.dismiss):
* UserInterface/Views/SourceCodeTextEditor.js:
(WI.SourceCodeTextEditor.prototype.dialogWasDismissed): Deleted.
Rename the handler when dismissing a dialog with a value so that
it is clear this is only on a successful value and not an invalid value.
Dialog now only calls this delegate with a valid value. This addresses
exceptions seen when the GoToLineDialog was dimissed via Escape and
an unexpected `null` value was being handled in the delegate.

* UserInterface/Views/GoToLineDialog.js:
(WI.GoToLineDialog.prototype._handleKeydownEvent):
* UserInterface/Views/OpenResourceDialog.js:
(WI.OpenResourceDialog.prototype._handleKeydownEvent):
Prevent default on the event whenever we dismiss the dialog.
Without this, the key event was also being sent to the editor.
In this case, that meant an "Enter" key was replacing the new
selection in the TextEditor with a newline!

Identifier: 205583@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237212 268f45cc-cd09-0410-ab3c-d52691b4dbfc