[CG] Adopt CG SPI for non-even cornered rounded rects https://bugs.webkit.org/show_bug.cgi?id=190155
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2018-10-18
Reviewed by Simon Fraser.

Source/WebCore:

Instead of creating bezier curves for the non-even corners of the rounded
rects, we should use the optimized SPI provided by CG.

* platform/graphics/cg/PathCG.cpp:
(WebCore::Path::platformAddPathForRoundedRect):

Source/WebCore/PAL:

* pal/spi/cg/CoreGraphicsSPI.h:

LayoutTests:

This test fails on iOS simulator because of just one pixel difference.
I think it happens because of anti aliasing the color at the border of
the black shadow. Since this test is testing the radius attribute of the
CSS box-shadow and this should not be affected by whether the shadow has
non-even rounded corners or not, I am going to change it to have even
rounded corners.

* fast/box-shadow/box-shadow-with-zero-radius-expected.html:
* fast/box-shadow/box-shadow-with-zero-radius.html:

Identifier: 205645@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@237280 268f45cc-cd09-0410-ab3c-d52691b4dbfc